+++ src/main/java/ntt/ntt_ms_accounts/dto/BankAccountRequestDto.java
@@ -3,10 +3,12 @@ package ntt.ntt_ms_accounts.dto;
 import lombok.Data;
 import ntt.ntt_ms_accounts.enums.AccountType;
 
+import java.math.BigDecimal;
+
 @Data
 public class BankAccountRequestDto {
     private String customerId;
     private AccountType accountType; // "SAVINGS", "CURRENT", "FIXED_TERM"
     private String accountNumber;           // númeroCuenta
-    private double balance;                 // saldo
+    private BigDecimal balance;                 // saldo
 }
